Move migration script to a dedicated scripts folder
